+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.9.0
+
+_Feat_
+
+- [Issue #55](https://github.com/XPEHO/xpeapp_admin/issues/55) Bulk import of QVST questions
+- Add info icon on the response reference widget
+
 ## 0.8.1
 
 _Fix_